Test all the things, injected or constructed.
